pythonproject/requirements.txt
allure_python_commons
clickhouse_sqlalchemy
Flask
jsonpath
pandas
paramiko
pymongo
PyMySQL
PyQt5
pytest
PyYAML
redis
requests
SQLAlchemy
urllib3
faker
xlrd==2.0.1
xlwt
xlutils
jenkins